Moved constants.py to src * Refactored imports from constants.py in sync.py
