+++ sync.py
@@ -1,5 +1,5 @@
 from src.client import Client
-from constants import OntologiesToSync, TablesToSync, TablesToDelete
+from src.constants import OntologiesToSync, TablesToSync, TablesToDelete
 from util import Util
